backend/requirements.txt
fastapi[standard]
sqlalchemy
python-dotenv
psycopg2-binary
PyJWT
passlib[bcrypt]
python-multipart
alembic